(fix) Move stores to settings root
