+++ strategy-builder-gui/src/pages/Settings/stores/counters.ts
