+++ strategy-builder-gui/src/pages/Settings/stores/selects.ts
